feat: add custom image model with uncroppable support
