+++ home/models.py
@@ -1,6 +1,8 @@
-# from django.db import models
+from django.db import models
 from wagtail.admin.panels import FieldPanel
 from wagtail.fields import RichTextField
+from wagtail.images.models import AbstractImage, AbstractRendition
+from wagtail.images.models import Image as WagtailImage
 from wagtail.models import Page
 
 
@@ -10,3 +12,23 @@ class HomePage(Page):
     content_panels = Page.content_panels + [
         FieldPanel("body"),
     ]
+
+
+class CustomImage(AbstractImage):
+    uncroppable = models.BooleanField(default=False)
+
+    admin_form_fields = WagtailImage.admin_form_fields + ("uncroppable",)
+
+    class Meta(AbstractImage.Meta):
+        verbose_name = "image"
+        verbose_name_plural = "images"
+        permissions = [
+            ("choose_image", "Can choose image"),
+        ]
+
+
+class CustomRendition(AbstractRendition):
+    image = models.ForeignKey(CustomImage, on_delete=models.CASCADE, related_name="renditions")
+
+    class Meta:
+        unique_together = (("image", "filter_spec", "focal_point_key"),)
