+++ tgno/settings/base.py
@@ -172,6 +172,7 @@ USE_L10N = True
 
 USE_TZ = True
 
+WAGTAILIMAGES_IMAGE_MODEL = "home.CustomImage"
 
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/4.2/howto/static-files/
